requirements.txt
numpy==1.18.1
colorlog==4.0.2
gym==0.16.0
h5py==2.9.0
ipdb==0.12.2
jupyter==1.0.0
moviepy==1.0.0
mpi4py==3.0.2
mujoco-py>=2.0.2.5
opencv-python==4.1.0.25
plotly==4.1.0
pygame==1.9.6
pyopengl==3.1.0
pyquaternion==0.9.5
scikit-image==0.15.0
seaborn==0.9.0
Pillow<7
torchvision==0.4.0
wandb==0.8.9
PyYAML==5.4
